% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlationNetwork_Function.R
\name{cor.network}
\alias{cor.network}
\title{Produces a network graph of the most correlated pairs of terms found throughout the corpus}
\usage{
cor.network(data.td, corlimit)
}
\arguments{
\item{data.td}{A tidy dataset}

\item{corlimit}{A user defined numeric correlation limit between 0 and 1}
}
\value{
A network graph of the most frequent bigrams down to a user defined \code{corlimit}
}
\description{
description
}
