% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corpusPrep_Function.R
\name{corpusPrep}
\alias{corpusPrep}
\title{Allows the user check and prepare their corpus to be in a usable state for the remainder of the commands}
\usage{
corpusPrep(data.td)
}
\arguments{
\item{data.td}{A tidy dataset}
}
\value{
Returns the same corpus, with appropriate column names, correct column classes, and prepped data
}
\description{
description
}
