% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/July_Articles.R
\docType{data}
\name{July_Articles}
\alias{July_Articles}
\title{July 1 - July 31, 2017 News Articles}
\format{An object of class \code{data.frame} with 22025 rows and 7 columns.}
\usage{
July_Articles
}
\description{
A clean dataset containing 22,025 news articles from nine 
   recognized news sources collected between July 01 - July 31, 2017. 
 
 \describe{ 
   \item{Title}{The title of each news article} 
   \item{Author}{Identifies the
   author of each news article. Unidentified authors are represented by
   "unknown" } 
   \item{Date}{Publication date of each news article given in the
   "YYYYMMDD HH:MM:SS" format} 
   \item{URL}{Weblink to each news article} 
   \item{NewsSource}{Represents the name of the news source where each
   article was published} 
   \item{Text}{Text from each news article} 
   \item{ArticleNo}{Sequential list of each article in the dataset} }
}
\keyword{datasets}
