% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keywordCorpus_Function.R
\name{keyword.corpus}
\alias{keyword.corpus}
\title{Create a subcorpus based on a user specified keyword}
\usage{
keyword.corpus(data.td, keyword)
}
\arguments{
\item{data.td}{A tidy dataset}

\item{keyword}{A user specified keyword}
}
\value{
Creates a subcorpus with only documents containing \code{keyword}
}
\description{
Create a reduced subcorpus of documents which contain a specified keyword.
}
\details{
This function will create a new subcorpus of documents, from the
  original corpus,  in which each document contains the user specified
  \code{keyword}. All other documents will be excluded from this subcorpus.
}
\examples{
\donttest{
hurricane <- keyword.corpus(Articles, "hurricane")
}
}
