% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/printNgrams_Function.R
\name{ngram.print}
\alias{ngram.print}
\title{Produces a bar graph of the most frequent n-grams found within a corpus}
\usage{
ngram.print(data.td, ngram = 1, num = 15)
}
\arguments{
\item{data.td}{A tidy dataset}

\item{ngram}{A user defined integer representing the number of tokens to analyze, 'n'}

\item{num}{A user definied number of most frequenct ngrams to plot in the bar graph}
}
\value{
A bar graph of the most frequent \code{ngram} found within a corpus
}
\description{
description
}
