% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readTabular.R
\name{readTabular}
\alias{readTabular}
\title{Reproduction of the readTabular from the tm package}
\usage{
readTabular(mapping)
}
\arguments{
\item{mapping}{A named list of characters. The constructed reader will map each character entry to the content or metadatum of the text document as specified by the named list entry. Valid names include content to access the document's content, and character strings which are mapped to metadata entries.}
}
\value{
Formally this function is a function generator, i.e., it returns a function (which reads in a text document) with a well-defined signature, but can access passed over arguments (e.g., the mapping) via lexical scoping.
}
\description{
Return a function which reads in a text document from a tabular data structure (like a data frame or a list matrix) with knowledge about its internal structure and possible available metadata as specified by a so-called mapping.
}
