% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runmyapp.R
\name{runmyapp}
\alias{runmyapp}
\title{Runs an app in my package}
\usage{
runmyapp(app_name, ...)
}
\arguments{
\item{app_name}{Character string for a directory in this package}

\item{...}{Additional options passed to shinyApp}
}
\value{
A printed shiny app
}
\description{
Runs as shiny app
}
\examples{
\dontrun{runmyapp('myfirstApp')
}
}
