% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topicNumber_Function.R
\name{topic.number}
\alias{topic.number}
\title{Produces a plot identifying the estimated correct number of topics to choose. Based on the `ldatuning` package.}
\usage{
topic.number(data.td, x = 0.9)
}
\arguments{
\item{data.td}{A tidy dataset}

\item{x}{A user defined term sparsity limit}
}
\value{
Plot indicating the estimated number of topics found within the corpus, \code{data.td}
}
\description{
description
}
