% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bigramNetwork_Function.R
\name{bigram.network}
\alias{bigram.network}
\title{Produces a network graph of most frequent bigrams found throughout the corpus}
\usage{
bigram.network(data.td, freq)
}
\arguments{
\item{data.td}{A tidy dataset}

\item{freq}{A user defined numeric vector containing an user defined limit of most frequent bigrams to include}
}
\value{
A network graph of the most frequent bigrams down to a user defined \code{freq}
}
\description{
description
}
