% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlationNetwork_Function.R
\name{cor.network}
\alias{cor.network}
\title{Network graph of the most correlated pairs of terms found throughout the corpus}
\usage{
cor.network(data.td, corlimit)
}
\arguments{
\item{data.td}{A tidy dataset}

\item{corlimit}{A user defined numeric correlation limit between 0 and 1}
}
\value{
A network graph of the most correlated word pairs down to a user defined \code{corlimit}
}
\description{
This function visualizes the most frequent pair of terms found 
  throughout a corpus of documents based on the Phi Correlation 
  Coefficient. The function will only produce word pairs at or above a
  provided correlation limit.
}
\details{
This function identifies all of the  terms found with in a corpus
  and sums the number of times each term is found within the same document
  relative to the other terms. With this information, each pair of terms is
  provided a correlation value using the Phi Coefficient. The function then generates
  an undirected network graph connecting each word (node) in a word pair.
  Correlated words are then connected by edges (lines) whose thickness varies based on
  the value of the correlation coefficient for each word pair.
}
\examples{
\donttest{
cor.network(Articles, 200)
}
}
