% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corpusPlot_Function.R
\name{corp.plot}
\alias{corp.plot}
\title{Produces a time series plot of the frequency of articles within a corpus by date}
\usage{
corp.plot(data.td)
}
\arguments{
\item{data.td}{A tidy dataset}
}
\value{
Time series line plot of documents in \code{data.td} by day
}
\description{
description
}
