% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iso_dateCorpus_Function.R
\name{iso.date}
\alias{iso.date}
\title{Creates a subcorpus based on user specified date ranges}
\usage{
iso.date(data.td, Beg.date, End.date)
}
\arguments{
\item{data.td}{A tidy dataset}

\item{Beg.date}{User specified beginning date}

\item{End.date}{User specified end date}
}
\value{
Creates a subcorpus containing only documents produced between \code{Beg.date} and \code{End.date}
}
\description{
description
}
