% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keywordCorpus_Function.R
\name{keyword.corpus}
\alias{keyword.corpus}
\title{Creates a subcorpus based on a user defined keyword}
\usage{
keyword.corpus(data.td, keyword)
}
\arguments{
\item{data.td}{A tidy dataset}

\item{keyword}{A user defined keyword}
}
\value{
Creates a subcorpus with only documents containing \code{keyword}
}
\description{
description
}
