% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keywordSearch_Function.R
\name{kwsearch}
\alias{kwsearch}
\title{Produces a time series plot of the frequency of articles, containing a user specified keyword, within a corpus by date}
\usage{
kwsearch(data.td, keyword)
}
\arguments{
\item{data.td}{A tidy dataset}

\item{keyword}{A user defined keyword}
}
\value{
Time series line plot of documents in \code{data.td}  which contain \code{keyword} by day
}
\description{
description
}
