% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topicModels_Function.R
\name{topic.models}
\alias{topic.models}
\title{Determines and classifies documents within a corpus to a user defined number of topics}
\usage{
topic.models(data.td, k = 16, x = 0.9)
}
\arguments{
\item{data.td}{A tidy dataset}

\item{k}{User defined number of topics}

\item{x}{A user defined term sparsity limit}
}
\value{
Attaches a new column to \code{data.td} identifying each document in a corpus as belonging to a specific topic
}
\description{
description
}
