% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WordAssocs_Function.R
\name{word.Assocs}
\alias{word.Assocs}
\title{Produces a time series plot of terms associated with a user defined word up to a specified correlation limit}
\usage{
word.Assocs(data.td, term, corlimit)
}
\arguments{
\item{data.td}{A tidy dataset}

\item{term}{A user defined character vector containing no spaces}

\item{corlimit}{A user defined numeric correlation limit between 0 and 1}
}
\value{
Time series line plot of words associated with \code{term}
}
\description{
description
}
